+++ extension/src/constants.ts
@@ -1,2 +1,3 @@
 export const componizerCommandName = 'extension.componize';
 export const defaultComponentName = 'NewComponent';
+export const defaultSkipImport = 'false';