+++ schematics/ng-componize/index.ts
@@ -1,34 +1,86 @@
-import { Rule, SchematicContext, Tree, chain, externalSchematic } from '@angular-devkit/schematics';
-import { Schema } from './schema';
-import { getProjectDetails, createTemplates, tryUpdateTemplate, getActiveFilePath, getNewComponentDetails } from './utils';
+import {
+  chain,
+  externalSchematic,
+  Rule,
+  SchematicContext,
+  Tree,
+} from '@angular-devkit/schematics';
 import { getWorkspace } from '@schematics/angular/utility/config';
+import { Schema } from './schema';
+import {
+  createTemplates,
+  getActiveFilePath,
+  getNewComponentDetails,
+  getProjectDetails,
+  logInFile,
+  tryUpdateTemplate,
+} from './utils';
 
 export function ngComponize(options: Schema): Rule {
+  overrideSkipImport(options);
   return chain([
     externalSchematic('@schematics/angular', 'component', options),
-    componize(options)
+    componize(options),
   ]);
 }
 
-
-function componize(options: Schema): (tree: Tree, _context: SchematicContext) => Tree {
+function componize(
+  options: Schema
+): (tree: Tree, _context: SchematicContext) => Tree {
   return (tree: Tree, _context: SchematicContext) => {
-    const { path: projectPath, prefix: projectPrefix } = getProjectDetails(getWorkspace(tree), options.project);
+    logInFile(JSON.stringify(options), options.debugMode);
+
+    const { path: projectPath, prefix: projectPrefix } = getProjectDetails(
+      getWorkspace(tree),
+      options.project
+    );
+
+    const {
+      selector: newComponentSelector,
+      directory: newComponentDir,
+    } = getNewComponentDetails(
+      projectPath,
+      options.name,
+      options.prefix || projectPrefix
+    );
 
     const activeFilePath = getActiveFilePath(projectPath, options.activeFile);
-    const { selector: newComponentSelector, directory: newComponentDir } = getNewComponentDetails(projectPath, options.name, options.prefix || projectPrefix);
 
     const activeFileBuffer = tree.get(activeFilePath);
-    if (activeFileBuffer) {
-      const { newTemplate, updatedTemplate } = createTemplates(activeFileBuffer, newComponentSelector, options.start, options.end);
-      tree.overwrite(activeFilePath, updatedTemplate);
 
-      tree.getDir(newComponentDir)
-        .visit(filePath => {
-          tryUpdateTemplate(filePath, () => tree.overwrite(filePath, newTemplate));
-        });
+    if (activeFileBuffer) {
+      const {
+        newComponentTemplate,
+        updatedComponentTemplate,
+        foundItem,
+      } = createTemplates(
+        activeFileBuffer.content.toString(),
+        newComponentSelector,
+        options.start,
+        options.end
+      );
+
+      tree.overwrite(activeFilePath, updatedComponentTemplate);
+
+      tree.getDir(newComponentDir).visit((filePath) => {
+        tryUpdateTemplate(
+          filePath,
+          () => tree.overwrite(filePath, newComponentTemplate),
+          tree,
+          foundItem,
+          options.name
+        );
+      });
     }
 
     return tree;
   };
+}
+
+function overrideSkipImport(options: Schema): void {
+  if (!!options.customSkipImport && options.customSkipImport === 'true') {
+    options.skipImport = true;
+  } else {
+    options.skipImport = false;
+  }
 }