+++ schematics/ng-componize/parser/binding-type-lookup.ts
@@ -1,0 +1,5 @@
+export enum BindingTypeLookup {
+  InterpolationLookup,
+  AttributeLookup,
+  AngularDirectiveLookup,
+}