+++ schematics/ng-componize/schema.ts
@@ -1,7 +1,9 @@
 import { Schema as ComponentOptions } from '@schematics/angular/component/schema';
 
 export interface Schema extends ComponentOptions {
-    activeFile: string;
-    start: number;
-    end: number;
+  activeFile: string;
+  start: number;
+  end: number;
+  customSkipImport: string;
+  debugMode: string;
 }