+++ extension/CHANGELOG.md
@@ -24,3 +24,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.0.3]
 
 - [Fix] Fixed wrong [slash after extraction](https://github.com/bilelmsekni/componizer/issues/5). Thanks to [Maykon Oliveira](https://github.com/maykon-oliveira)
+
+## [0.0.4]
+
+- [Feature] add skipImport question
+- [Feature] add hidden debugMode
+- [Feature] handle 3 types of bindings : attribute binding, two-way bindings, angular directive. new dependency : ts-morph  to manipulate new component typescript
+- [Enhancement] update mochajs and add watch task
+- [Enhancement] local install extension/schematic task
+- Many thanks to [Hugo Mercier](https://github.com/hugoparis19) and [Alexandre Crochet](https://github.com/ekaliroots)